requirements.txt
Flask
sqlalchemy
python-dotenv
flask-cors
bcrypt
pyjwt
mysql-connector-python